refactor to separate files ready for ProxyChangeset
