+++ src/-private/utils/consts.ts
@@ -1,0 +1,4 @@
+export const { keys } = Object;
+export const defaultValidatorFn = () => true;
+export const defaultOptions = { skipValidate: false };
+export const DEBUG = process.env.NODE_ENV !== 'production';
