+++ src/-private/utils/maybe-unwrap-proxy.ts
@@ -1,0 +1,5 @@
+import { Content } from '../../types';
+
+export default function maybeUnwrapProxy(content: Content): any {
+  return content;
+}
